Add `xdg-desktop-portal` to KDE packages.
According to [this post](https://phabricator.kde.org/T10189), the `xdg-desktop-portal-kde` is simply a plugin for the `xdg-desktop-portal` package. This helps us to get KDE dialogs instead of the GTK3 ones on many apps such as Firefox.

This also needs the `GTK_USE_PORTAL=1` environment variable to be set to use KDE dialogs.