fix info message for legacy command
Refers to archiso commit 2e1ddec0:
Change the way _show_config() displays information about the build environment,
as displaying a profile directory as a legacy command is confusing.
The function now prints a deprecation message if '$command_name' is not a
directory (i.e. not a profile). Remove 'command_' prefix from help output for
legacy commands (the prefix is only in use for internal functions).