set CacheDir and HookDir for profile more sanely
Refers to archiso commit f3af5692:
mkaui:
Change `_pacman()` to use the modified pacman.conf from the work_dir, instead
of using the unmodified pacman.conf.

Change `_make_pacman_conf()` to compare the system's and software's CacheDir
setting and use the software's CacheDir setting only if it's not the default
and not the same as the system's.

Always set the HookDir to the airootfs override directory, so that no hooks
from the host system are being run.

Remove DBPath, LogFile and RootDir settings from the work_dir pacman.conf as
they are otherwise referring to the host system, even if pacman is being
called with the `-r` flag.

Fix a typo in _make_custom_airootfs().