support EROFS
Refers to archiso commit bc67933a:
EROFS, like Squashfs, is a read-only file system. It can be used to
store airootfs in an image file.
Its advantage is the support for POSIX ACLs. EROFS downside is that
currently it only supports LZ4 compression (LZMA support is not yet
fully implemented).

A difference from Squashfs is that, EROFS stores change time (ctime) not
modification time (mtime). The reverse is true for Squashfs.