force PGP signature file extension
Refers to archiso commit cc735dbb:
Force the file extension in use for the PGP signatures of the rootfs to
always be .sig.
When gnupg's 'armor' configuration option is used, the output otherwise
defaults to using .asc.
As the verification hook in mkinitcpio-archiso expects the .sig file
extension, verifying the rootfs will fail in that scenario.