set more generic output for signatures
Refers to archiso commit 42cdf867:
Change the help output to reflect that the `-g` option is generically
signing a rootfs (which may be e.g. squashfs or erofs).
Change the output of `_mksignature()` to be more generic, as it signs
any type of understood rootfs image (which may be e.g. squashfs or erofs).