add buildmode to export netboot artifacts
Refers to archiso commit d54bf635:
Implement a buildmode to export artifacts required for netboot with
IPXE.

When providing the buildmode 'netboot' via profiledef.sh or the `-m`
option, all targets necessary to create an ISO
medium are built, but the components required for netboot are exported
to the output dir.

Optionally, it is possible to provide a set of certificates for
codsigning using the `-c` option, where the first file
is considered as the signer certificate and the second as the key.

Add `_export_netboot_artifacts()` to copy build artifacts to the output
directory.

Add `_sign_netboot_artifacts()` to codesign the netboot artifacts in the
work directory.

Add `_validate_requirements_buildmode_netboot()` to check for openssl.

Add `_build_iso_base()` to implement common function calls between the
'iso' and the 'netboot' buildmodes.

Add `_build_buildmode_netboot()` to make use of `_build_iso_base()`,
(optionally) `_sign_netboot_artifacts()` and
`_export_netboot_artifacts()`.

Change `_build_buildmode_iso()` to make use of `_build_iso_base()`.

Add `-c` as an option to mkarchiso to read in a list of file names.

Unify the output of `_usage()` by using the same definition style for
lists of strings provided to options that accept
them (e.g. `-c`, `-m`, `-p`).