syslinux PXE: Forced all TFTP paths to be absolute
Refers to archiso commit 0c6ecb6b:
By default, syslinux interprets TFTP paths as relative to the location
of the *.c32 modules, regardless whether a path starts with a slash.
Without the `::` that I added to all of these paths, syslinux cannot
find these files. It searches for them in
/%INSTALL_DIR%/syslinux//%INSTALL_DIR%/boot/x86_64/vmlinuz-linux, etc.