aui-mkiso: do not try to use an non existent GPG public key file
archiso commit d31f3884:
The `bootstrap` build mode never calls `_export_gpg_publickey`, so even
if the GPG key is passed with the `-g` option and thus the `gpg_key`
variable is set, the `${work_dir}/pubkey.gpg` file will not exist.
This has not caused any issue so far because the `ARCHISO_GNUPG_FD` file
descriptor opens the file for both reading and writing, which means the
file gets created if it does not exist.

Assign the exported public key file name to a `gpg_publickey` variable
in `_export_gpg_publickey` and check for it when the file is used.

Since the exist status of the gpg command cannot be checked, look for
the exported public key file instead.