aui-mkiso: increase the additional free space added to efiboot.img
archiso commit 1ae3ff6f:
Less than 1 MiB (but rounded up to one) is needed to account for file
anddirectory entries, especially when using FAT32. Otherwise the ISO build
could stop with:

    Disk full

The rest of the increase is to allow more space for adding custom files
after the fact, i.e. when repacking the ISO. E.g. for the purposes of
adding a signed boot loader for Secure Boot or similar. See
https://wiki.archlinux.org/title/Unified_Extensible_Firmware_Interface/Secure_Boot#ISO_repacking