.automated_script.sh: add proper HTTP status checking
archiso commit 75e57093:
When a download attempt fails with an HTTP error, the body of the failed
request will be written to the output file with a 0 exit status, causing
the
execution of an error message as a script.

This issue is resolved by adding the --fail option to curl.