zfs-utils: disable use of libunwind
If libunwind is detected during building, it will automatically be
linked. Since this is more relevant for people doing
development/debugging and probably not for the majority of users of this
package, we are disabling it.
Commit 	4d72ad9d496a3a705f84169dc857d143e25c8d22 from AUR zfs-utils